e2e tests: made execution order stricter to prevent errors due to parallelism
Changelog: None
