+++ tests/e2e_tests/integration/05-deviceDetails.ts
